Extracted reusable functions
