+++ cypress/e2e/helpers/index.ts
@@ -1,1 +1,3 @@
 export { login, logout } from './login';
+export { selectPublishedEntities, selectRestrictedEntities } from './entitiesFilters';
+export { createUser } from './users';
